Move fixtures out of command tests into separate conftest for commands
